+++ lib/src/common/constants.dart
@@ -1,0 +1,8 @@
+/// [Constants] is a class for holding constant values.
+class Constants {
+  /// [isDarkKey] is a String for holding the `is_dark_key`.
+  static const String isDarkKey = "is_dark_key";
+
+  /// [themeBox] is a String for holding the `theme_box`.
+  static const String themeBox = "theme_box";
+}