+++ lib/src/common/theme_mode_builder_config.dart
@@ -1,4 +1,5 @@
 import 'package:hive/hive.dart';
+import 'package:theme_mode_builder/src/common/constants.dart';
 import 'package:theme_mode_builder/src/services/hive_storage_service.dart';
 import 'package:theme_mode_builder/src/theme_mode_builder/theme_mode_builder.dart';
 
@@ -18,11 +19,11 @@ class ThemeModeBuilderConfig {
   /// * false (light)
   static bool isDarkTheme() {
     final Box<bool> themeBox = HiveStorageService().getBox<bool>(
-      boxName: "theme_box",
+      boxName: Constants.themeBox,
     );
 
     final bool? isDarkTheme = themeBox.get(
-      "is_dark_key",
+      Constants.isDarkKey,
     );
 
     return isDarkTheme ?? false;
@@ -35,11 +36,11 @@ class ThemeModeBuilderConfig {
   /// * light (false)
   static Future<void> toggleTheme() async {
     final Box<bool> themeBox = HiveStorageService().getBox<bool>(
-      boxName: "theme_box",
+      boxName: Constants.themeBox,
     );
 
     await themeBox.put(
-      "is_dark_key",
+      Constants.isDarkKey,
       !isDarkTheme(),
     );
   }